Updated to fix wiwwiwi diagnostic_federator queue processing label bug. Custom merge into /main/branches/dev/toddjones/r7304_tvfff: r7523 cf. r7522 of /main/branches/dev/toddjones/r7137_wiwwiwi@7524
git-svn-id: https://code.metoffice.gov.uk/svn/monc/main/branches/dev/toddjones/r7304_tvfff@7525 0f676ef4-b20c-4647-9485-21614760d15f

+++ io/src/writers/writer_field_manager.F90
@@ -320,7 +320,7 @@ contains
     else
       generic=>generate_value_container(field_name, field_namespace, field_values, timestep, time, frequency, source)
       if (time_basis) then
-        call c_put_generic(field_ordering%timestep_to_value, conv_to_string(time), generic, .false.)
+        call c_put_generic(field_ordering%timestep_to_value, conv_to_string(nint(time)), generic, .false.)
       else
         call c_put_generic(field_ordering%timestep_to_value, conv_to_string(timestep), generic, .false.)
       end if
